move claim-incident-types to types dir
